Add delete before deploy setting (#864)

+++ src/tree/blob/BlobContainerTreeItem.ts
@@ -15,12 +15,13 @@ import { AzureStorageFS } from '../../AzureStorageFS';
 import { createAzCopyLocalLocation, createAzCopyRemoteLocation } from '../../commands/azCopy/azCopyLocations';
 import { azCopyTransfer } from '../../commands/azCopy/azCopyTransfer';
 import { IExistingFileContext } from '../../commands/uploadFiles';
-import { getResourcesPath, NotificationProgress, staticWebsiteContainerName } from "../../constants";
+import { configurationSettingsKeys, getResourcesPath, NotificationProgress, staticWebsiteContainerName } from "../../constants";
 import { ext } from "../../extensionVariables";
 import { TransferProgress } from '../../TransferProgress';
 import { createBlobContainerClient, createChildAsNewBlockBlob, IBlobContainerCreateChildContext, loadMoreBlobChildren } from '../../utils/blobUtils';
 import { throwIfCanceled } from '../../utils/errorUtils';
 import { localize } from '../../utils/localize';
+import { getWorkspaceSetting } from '../../utils/settingsUtils';
 import { getUploadingMessageWithSource, uploadLocalFolder } from '../../utils/uploadUtils';
 import { ICopyUrl } from '../ICopyUrl';
 import { IStorageRoot } from "../IStorageRoot";
@@ -207,6 +208,7 @@ export class BlobContainerTreeItem extends AzureParentTreeItem<IStorageRoot> imp
         notificationProgress: NotificationProgress,
         cancellationToken: vscode.CancellationToken
     ): Promise<string> {
+        ext.outputChannel.appendLog(localize('deploying', 'Deploying to static website "{0}/{1}"', this.root.storageAccountId, this.container.name));
         const retries: number = 4;
         await retry(
             async (currentAttempt) => {
@@ -216,28 +218,28 @@ export class BlobContainerTreeItem extends AzureParentTreeItem<IStorageRoot> imp
                     ext.outputChannel.appendLog(message);
                 }
 
-                // Find existing blobs
-                let blobsToDelete: azureStorageBlob.BlobItem[] = [];
-                blobsToDelete = await this.listAllBlobs(cancellationToken);
+                if (getWorkspaceSetting<boolean>(configurationSettingsKeys.deleteBeforeDeploy)) {
+                    // Find existing blobs
+                    let blobsToDelete: azureStorageBlob.BlobItem[] = [];
+                    blobsToDelete = await this.listAllBlobs(cancellationToken);
 
-                if (blobsToDelete.length) {
-                    let message = `The storage container "${this.friendlyContainerName}" contains ${blobsToDelete.length} files. Deploying will delete all of these existing files.  Continue?`;
-                    let deleteAndDeploy: vscode.MessageItem = { title: 'Delete and Deploy' };
-                    const result = await vscode.window.showWarningMessage(message, { modal: true }, deleteAndDeploy, DialogResponses.cancel);
-                    if (result !== deleteAndDeploy) {
-                        context.telemetry.properties.cancelStep = 'AreYouSureYouWantToDeleteExistingBlobs';
-                        throw new UserCancelledError();
+                    if (blobsToDelete.length) {
+                        let message = `The storage container "${this.friendlyContainerName}" contains ${blobsToDelete.length} files. Deploying will delete all of these existing files.  Continue?`;
+                        let deleteAndDeploy: vscode.MessageItem = { title: 'Delete and Deploy' };
+                        const result = await vscode.window.showWarningMessage(message, { modal: true }, deleteAndDeploy, DialogResponses.cancel);
+                        if (result !== deleteAndDeploy) {
+                            context.telemetry.properties.cancelStep = 'AreYouSureYouWantToDeleteExistingBlobs';
+                            throw new UserCancelledError();
+                        }
                     }
+
+                    // Delete existing blobs
+                    let transferProgress = new TransferProgress('blobs', blobsToDelete.length, 'Deleting');
+                    await this.deleteBlobs(blobsToDelete, transferProgress, notificationProgress, cancellationToken, context.telemetry.properties);
+
+                    // Reset notification progress. Otherwise the progress bar will remain full when uploading blobs
+                    notificationProgress.report({ increment: -1 });
                 }
-
-                ext.outputChannel.appendLog(`Deploying to static website ${this.root.storageAccountName}/${this.container.name}`);
-
-                // Delete existing blobs (if requested)
-                let transferProgress = new TransferProgress('blobs', blobsToDelete.length, 'Deleting');
-                await this.deleteBlobs(blobsToDelete, transferProgress, notificationProgress, cancellationToken, context.telemetry.properties);
-
-                // Reset notification progress. Otherwise the progress bar will remain full when uploading blobs
-                notificationProgress.report({ increment: -1 });
 
                 // Upload files as blobs
                 await uploadLocalFolder(context, this, sourceFolderPath, destBlobFolder, notificationProgress, cancellationToken, 'Uploading', false);
